requirements.txt
tensorflow-gpu==2.0.0
numpy == 1.18.1
scipy=1.3.1
